Authorization and authentication using jwt

+++ controllers/api/v1/posts_api.js
@@ -1,6 +1,5 @@
 const Post = require("../../../models/post");
 const Comment = require("../../../models/comment");
-
 module.exports.index = async function (req, res) {
   let posts = await Post.find({})
     .sort("-createdAt")
@@ -22,17 +21,23 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "Post and associated comments deleted!",
-    });
+      return res.json(200, {
+        message: "Post and associated comments deleted successfully!",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post!",
+      });
+    }
   } catch (err) {
-    console.log(err);
+    console.log("********", err);
     return res.json(500, {
-      message: "Internal Server error",
+      message: "Internal Server Error",
     });
   }
 };
